+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0 (unreleased)
 
 - implement validation for models
+- wrap `sqlalchemy.orm.relationship` with the configured `Query` class
 
 ## v0.1.0 (2018/09/24)